Score change + minor clue edit
Changed the score to starting back at 0, and the team gets 10 points for
completing the game (i.e., getting to the "congratulations!" screen.

Also fixed the layout so that it shows 0 as the initial team score.

Shortened one of the clues so all of the text would fit when using up
all clues.